move app settings getter to settings controller static class
